requirements.txt
bcc==0.18.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
Brlapi==0.8.3
bs4==0.0.2
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
Flask==3.0.1
Flask-Bcrypt==1.0.1
fonttools==4.49.0
gunicorn==21.2.0
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.3
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==2.1.4
matplotlib==3.8.3
more-itertools==8.10.0
moviepy==1.0.3
mysql-connector==2.2.9
netaddr==0.8.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
packaging==23.2
passlib==1.7.4
pillow==10.2.0
pop-transition==1.1.2
proglog==0.1.10
protobuf==3.12.4
psycopg2-binary==2.9.9
pycairo==1.20.1
pycups==2.0.1
pydbus==0.6.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
repolib==2.2.1
repoman==1.4.0
requests==2.25.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
soupsieve==2.5
systemd-python==234
tqdm==4.66.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
